refactor(estimation): extract covariance/SE subsystem into estimation::covariance + dedup the 8-way cov-step block
Round-3 safe-tier modularization (T4). Two behavior-preserving changes:

1. Pure code motion: the entire covariance/standard-error subsystem (~1093 LOC:
   CovarianceOutput/CovarianceStepResult, compute_covariance, invert_psd_with_floor
   + RegularizedInverse, the FD-step selector, score cross-product, combine_covariance,
   the non-PD SIR fallback, eigen extraction, and the progress reporter) moves verbatim
   out of outer_optimizer.rs into a new src/estimation/covariance.rs. outer_optimizer
   retains a `pub(crate) use` facade so every historical
   `outer_optimizer::{compute_covariance, CovarianceStepResult}` import path (run_covariance.rs
   et al.) keeps working. 12 private helpers are widened to pub(crate) purely so the cov
   unit tests (kept in outer_optimizer's test mod) reach them across the boundary.

2. Dedup: the gated `run_covariance_step && !is_cancelled` timer + Success/Unusable/FailedNonPd
   match was copy-pasted (byte-identical apart from args + one verbose string) into 8 estimator
   finalizers (outer_optimizer eval/nlopt/bfgs, gauss_newton pure/hybrid, impmap, saem,
   trust_region). Replaced by one `run_covariance_step(..) -> CovStepOutcome` helper that each
   site calls, folding its own verbose flag in via `.then_some(msg)`. The helper contains ZERO
   floating-point arithmetic, so it cannot change any numeric result; the verbose stderr line
   and warning-vector ordering are preserved at the same program points.

Net ~-190 LOC. The only micro-change is pack_params hoisted just outside the gate in
impmap/saem (pure, unused-when-skipped) — behavior-preserving.

Verified: cargo test --profile ci-test --lib => 2505 passed, 0 failed. The load-bearing
anchors pass unchanged: test_compute_covariance_reconverged_matches_scalar_fd_with_factor_two
(pins each SE to scalar-FD), test_combine_covariance_sandwich_matches_explicit_product,
test_invert_psd_with_floor_*, test_assemble_score_cross_product_cancelled, cov_progress math.
