+++ src/estimation/impmap.rs
@@ -35,9 +35,7 @@ use crate::estimation::importance_sampling::{
     compute_posterior_hessian, find_optimal_iscale, subject_is_draws, SubjectDraws,
 };
 use crate::estimation::inner_optimizer::{find_ebe, EbeResult, InnerLoopStats};
-use crate::estimation::outer_optimizer::{
-    compute_covariance, pop_nll, CovarianceStepResult, OuterResult,
-};
+use crate::estimation::outer_optimizer::{pop_nll, OuterResult};
 use crate::estimation::parameterization::{compute_mu_k, pack_params, theta_packs_log};
 use crate::estimation::saem::{floor_omega_diagonal, get_mu_ref_pairs};
 use crate::pk::EventPkParams;
@@ -1188,42 +1186,22 @@ fn run_mcem(
         );
 
     // ---- Covariance step ----
-    let mut sir_fallback_proposal: Option<DMatrix<f64>> = None;
-    let (covariance_matrix, covariance_wall_time_secs) =
-        if options.run_covariance_step && !crate::cancel::is_cancelled(&options.cancel) {
-            let cov_timer = std::time::Instant::now();
-            let packed = pack_params(&final_params);
-            let cm = match compute_covariance(
-                &packed,
-                &final_params,
-                model,
-                population,
-                &eta_hats,
-                &h_matrices,
-                &final_kappas,
-                options,
-            ) {
-                CovarianceStepResult::Success(out) => {
-                    warnings.extend(out.warnings);
-                    Some(out.matrix)
-                }
-                CovarianceStepResult::Unusable(msg) => {
-                    warnings.push(msg);
-                    None
-                }
-                CovarianceStepResult::FailedNonPd {
-                    reason,
-                    fallback_proposal,
-                } => {
-                    warnings.push(reason);
-                    sir_fallback_proposal = Some(fallback_proposal);
-                    None
-                }
-            };
-            (cm, cov_timer.elapsed().as_secs_f64())
-        } else {
-            (None, 0.0)
-        };
+    let packed = pack_params(&final_params);
+    let out = crate::estimation::covariance::run_covariance_step(
+        &packed,
+        &final_params,
+        model,
+        population,
+        &eta_hats,
+        &h_matrices,
+        &final_kappas,
+        options,
+        None,
+    );
+    let covariance_matrix = out.matrix;
+    let covariance_wall_time_secs = out.wall_time_secs;
+    warnings.extend(out.warnings);
+    let sir_fallback_proposal = out.sir_fallback_proposal;
 
     // ---- Finalize trace ----
     let impmap_trace = if collect_trace {
